delete views2
delete views2
